requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dateutil==2.8.2
PyJWT==2.8.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
Faker==19.3.0
reportlab==4.0.4
openpyxl==3.1.2
Flask-CORS==4.0.0
gunicorn==21.2.0
